+++ packages/gestalt-codemods/1.8.0-1.9.0/__testfixtures__/heading-size-replace-md.output.js
@@ -1,0 +1,11 @@
+// @flow
+import React from 'react';
+import { Heading } from 'gestalt';
+
+export default function MDHeading() {
+  return (
+    <Heading>
+      md
+    </Heading>
+  );
+}