+++ packages/gestalt-codemods/1.8.0-1.9.0/__testfixtures__/text-size-replace-lg.output.js
@@ -1,0 +1,11 @@
+// @flow
+import React from 'react';
+import { Text } from 'gestalt';
+
+export default function LGText() {
+  return (
+    <Text>
+      lg
+    </Text>
+  );
+}