+++ packages/gestalt-codemods/1.8.0-1.9.0/__tests__/text-size-replace.test.js
@@ -1,0 +1,19 @@
+import { defineTest } from 'jscodeshift/dist/testUtils.js';
+
+jest.mock('../text-size-replace', () => {
+  return Object.assign(require.requireActual('../text-size-replace'), {
+    parser: 'flow',
+  });
+});
+
+describe('text-size-replace', () => {
+  [
+    'text-size-replace-xs',
+    'text-size-replace-sm',
+    'text-size-replace-md',
+    'text-size-replace-lg',
+    'text-size-replace-xl',
+  ].forEach(test => {
+    defineTest(__dirname, 'text-size-replace', { quote: 'single' }, test);
+  });
+});