+++ packages/gestalt/src/VideoControls.js
@@ -145,7 +145,7 @@ class VideoControls extends React.Component<Props> {
           </Touchable>
         </Box>
         <Box width={50} padding={2}>
-          <Text color="white" align="right" size="xs">
+          <Text color="white" align="right" size="sm">
             {timeToString(currentTime)}
           </Text>
         </Box>
@@ -159,7 +159,7 @@ class VideoControls extends React.Component<Props> {
           />
         </Box>
         <Box width={50} padding={2}>
-          <Text color="white" align="right" size="xs">
+          <Text color="white" align="right" size="sm">
             {timeToString(duration)}
           </Text>
         </Box>